Change prices of item receipt to Maybe
We need to modelize that at least of one of the price is known.
At this stage we it's easier to keep everything as Maybe.
We need to know the TAX rates to be able check if the price
combination is valid or not.  This will come with the next step.