Check locations in Csv matches existing location.
We actually use Glob, allowing location in the spreadsheet to be
glob pattern. However, this is slow it would be better to *expand*
loc pattern to a list of locations and check if they belongs
to the list. The main difference is globing allows to guess location
whereas expanding forces the user to be explicit.

For example, A[123], expands to A1 A2 A3, whereas
A? matches A1 A2 A3 even if A3 doesn't exists.