Remove "style completion" and fix date in stock adjustment display.
When doing a stocktake, varations not found are not entered. We are not
aware of the 0. In order to adjust the stock correctly we use to
generate a stock adjustment for variations not present within a style.

For example, we have two variations (Black and Red) of "T-Shirt".
We don't have any red left, so the stocktake only reports N black.
We haven't found any red, therefore they don't appears on the stocktake
even though the absence of it, means 0.

The previous code, tried to create an adjustment for the "missing"
variations. We don't do it anymore. The missing stocktakes needs
to be created when uploading the stocktake (using the complete style
options).