Use first detail reference as main reference
The Order reference which is stored for each detail is not part of
the spreadsheet details row but specfied as a order row instead.
It as the advantage of specifying it only once per file if only one
reference is used (which is the general case). Also that's how our
suppliers send it to us.

However, it has the disadvantage that we need to use the same form when
editing details. This is fine but it needs a starting reference is
nothing has been  providing in the cart. We use the reference of the
first details.