Add stocktake history
The barcode of a stocktake can be used more than once
when doing different stocktake (if the box hasn't been open and therefore
its content hasn't been changed). Because the unique constraint on stocktake
barcode (to avoid same barcode to be used twice and detect eventual mistake)
Successive stocktake of the same content (ie same barecode) can't be recorded.
The new history field allows to keep track on all of the scans.

Only new scans are push to the history list. Amendments are not stored as
considered as irrelevant.