Replace ZeroTake with QuickTake
QuickTake means we enter the quantity of a box,
not necessarily all of a style, resulting in a partial quantity on hand.
However we hope to "resolve" the discrepency between the real quantity
and the quantity entered by applying a modulo when creating the stock
adjustement.

For example a previous stock take gives 4 boxes of 6 pieces for a total of 24.
The quantity on hand in FA is (after a few items have been sold) 23.
After a quick check we notice that one box only contains 4 (instead of 5, as one
item has been sold.).
We can therefore deduce that we lost 1 items and adjust the stock accordingly
to 22 (one sold + one lost).