Display tabs ...
Not great, but will do for the moment.
The main problem is that we need to send the post data when changing tabs.
For this, we have to do some tricks and transform links to post event.
It might have been better to use a hidden field with the view mode,
So that resubmitting the form using another button should keep the current tab.