Add DC (commerce) entities.
To access tables from a external database
we use federated engine.
the federated tables are created by doing find and replace on
a create tables script (mysqlumd -d).
However this way doesn't allow to easily modify the table name.
by default federated table must have the same name as the original table.

We create all this table in a separate database (commerceX) and then generate some view
in the fa database (with a `dcx_` prefix) pointig to the federeated table.