Factorize update boxtake function to update history.
Also disable invalidating boxes when doing boxtake.
This is because we are not sure now what we want.
Even though it is legitimate to invalidate all boxes of a style
when a stocktake of 0 has been done. This means that we
can have the current states of the boxes changing when we don't want to
(planning for delivery for example). Also, boxes can still be there
(even with an ST barcode) if they have been used for something of a different
style.