Add Attributes to Roles. Might need to be reverted
I'm not sure it is really a good idea.
It allows a user to have different roles depending on the point of view
as all of it permissions can be filtered via attributes.
For example it allows Payroll/amount to work only for view using #view.
However, we are not sure it works well for other permissions
and doesn't really solve the initial problem with was to be able
to know easily what permission was missing when trying to unlock an item.
The old solution will create an artificial privilige "p#view" whereas displaying
a requirement for "p". But the new system does the same.