Refactoring: remove rank from NMapKey
NMapKey design was flawed as it was using the rank as a key,
meaning different key with the same rank (example different period)
will be seen at the same key.

Example a product being only sold in 2017 will have rank 1 For 2017
It will then appears in the 2018 column (2018 as rank 1).