Fix JS performance issue
Before, we would find all targets once before setting the onChange.
The advantage would be that the actual callbacks we only search
for the targets once. However, if there are many rows, it is really
slow to do at the startup so we moved the target selection in the callback
itself.