Payroll: clear timesheet instead of rejecting it when linked.
If a timesheet is link to some transaction in FA (via TransactionMap non Voided).
Reject will clear the timesheet item and payroll but leave the timesheet
instead of deleting everything. This allow to modify a timesheet
with payment already paid (which therefor have to stay in FA).
The amended timesheet will keep the link to the original one
and a new payment can enter with the correct amount and date using the
GUI.