Bank: sort and rebalance FA transactions
Statements in reconciliate sort transaction in the same order
as what is on the statement. When there is no statement, we
use instead the original FA position (within a reconcilited group).
Also, in both case, FA balance is recalculated to use the display order
order instead of original FA one.