TaxReport: Allow null customer category in ECSL
As ECSL uses customer categories, we need them to be up to date so
that all customer have their category computed.
To be sure to not miss any, we change the code so that null ECSl category
would be an error, forcing the category configuration to set "" instead.
This is not really convenient, so we are back to allow null.
All dodgy customer should be caught by the boxes themselves.

Another solution would be to recompute the category for new customers before
collecting it.