Report:@WIP Attempt to allow D&D between Browser and external application.
Doesn't work.
Firefox: D&D work from an external application but not from itself (which
happen if we drag a rupture panel to a text field back and forth : in that case
not dragevent is initiated and e.dataTransfer is null). But seems
to work from an external appl.
No DragEnter/leave are generated though, so the icon doesn't change.
Chrome: getData doesn't allways work even though we set it properly.