Items: Change base variations to base candidates
Now this field is there to give a list of base candidates.
Instead of chosing the first variations as a base, the first available candidate
is chosen (if possible).

However, if a base has been selected, it gets priority. This can be a problem
if the selected base hasn't been selected by the user but is just the first
available. Setting the base candidates and reloading wouldn't not the change
base  : the checked one will be selected.

To solve this problem, there is now a new "Clear/Set Variation" button which
clear all user variations settings (checkboxes and base, but not columns).