Planner: Fix error on parsing Date
File without header should be parsed first, otherwise a file with a single line
and no header would be seen as a empty file with ANY header.
However, this only works if, when parsing header (seen as a normal line) the
parsing fails without crashing (which was the can case when reading date).