Fix bug opening balance in reconciliate page
The initial balance used for FA was wrong.
It should work now, but maybe not.
The proble is due to transaction being in FA before the start date
but matched with an statement afte the start date.
We need all the transaction in between to calculate the FA balance
correctly. If this doesn't work the simplest is to recalculate the FA
balance first (after reordering per reconciliate group) and then
filtered the undesired group.