Replace is_last field by summary table
As a summary we need the last 0_stock_moves.trans_id and the last 0_gl_trans.counter
so that we can skip moves and transactions before that.
Looking at the last transaction doesn't work because a transaction might
only have a move_id OR and gl_detail.
By saving the summary in a different table, we keep the last used one,
which solves the problem.