Use audit trail sequence id to sort and group transaction.
Transaction needs to be ordered by date and within a day
in the order they actually have been entered in the system.
Transaction number is not enough as transaction can have different type.
However, all transactions have an entry in the audit table. The id
being sequencial, we can use use to find the order of the transaction
within a day.