Remove stock move pairs when no GL
This prevent the bug occuring when a GRN has been removed but
there is a transaction on the same day (or not)
In that case, the previous behavior would split the stock move pair
(which is fine) but then will try to match with the next invoice
resulting invoice and GRN being mistmatch.
Another way is to allow negative GRN to match GRN
the commit message for your changes. Lines starting