ItemCost: Fix update summary of item with closingDate
The problem is that the fixes can be on date which is greater
than the closingDate (last date that should in theory been looked at).

collectCostTransactions is modified to allow collecting on "todays" date
The problem is now the "openingDate" of the item with then new account.
We need to use the closingDate instead of the last transaction date
(which can be past the closingDay once we called fixGLBalance)