fix(itemCost): Moves on fixing date being loaded my mistake
Items which have a closing date (because their account has changed in
time) have a closing date. When fixing the gl balance we need
to collect and add to the summary the new gl trans created by the
fixing.
This transaction is normally after the closing date and therefore should
be filtered and so impossible to collect. To remedy at that we load
also transactions on the fixing day (if they occurs after the last seq
number). If some stock moves appears on that day (but on the new
account), they are loaded but don't have a corresponding gl trans (which
belongs to the other account). This generate an error resulting
in the summary not being updated correctly.

To solve this, we only load transaction and ignores moves when
collect for refresh.