Planner: Add 'soft' route to image
When the scenario is a unmodified scenario from a file,
the link to an image, instead of being a hard link to cached scenario,
is a link to a image of the scenario "by name".
This means that reloading an image page after modifying the file
describing the scenario will update the image.
Before the image was "baked".