ItemCost: Fix check in the past
not working if something in the past has been overriden.
This happen if we collect, fix, modify in the past (before fixing date)
collect again and fix.
Before fixing the 2nd time we should have checked for transaction in the past
but we didn't. This transaction is then buried by the 2nd fix.
This fix should detect it.