Planner(Refactor): Rework tiling to be lazy
reall small boxes and real big shelf (like error or pending)
can result in millions of slots. The curret box break  algo
can't keep up with it.

This commit replaces list of slots with breaks with a Slot, Slice
and Slices types. Those types should be able to be merged and sorted
lazily and therefore cope with millions of slots (if not used).