Add auto vars to index
Variations which don't exist in a "In filter" in the sku section
of the index (comming from an incomming purchase order for example)
will be displayed as missing.

In that case the "variations" used don't do a cross products (as it is
the case when using the variations fields) but only the variations
corresponding to a given sku will be seen as missing.