Planner: Fix copyWarehouse changing order of boxes
by removing it alltogether.
The problem we solves in that copyWarehouse would change the order in
which boxes are stored it Warehouse (the boxes field).
This is order is important because it is default order boxes are
processed (and sorted).
The current implementation of `copyWarehouse` copy shelves and boxes
belonging to shelves.
This means that boxes are copied in shelf order so boxes in a *early*
shelf are moved at the front of the boxes.
Also there is the problem that shelves store boxes in reverse order.
(which should be changed but that doesn't fix the fundamentol problem
which is shelf order)