BoxAdjustment : use batch and fix scan
Batch number is now used when sorting boxes to decide which box are
alive or not (we keep last batch if possible)

Scanning a box if active update its date (= last seen date).
Fix a bug when if two boxes of the same SKU are scanned let say A and B
where A is older than B but has been moved since the last scan.
Scanning will update A's date to today (but not B which hasn't changed).
Later on, A will be seen as *newer* than B (even though both have been
seen on the same day) and will result (when only once box is left) to
deactivate B instead of A.